pyproject.toml
[tool.poetry]
name = "product-watch"
version = "0.1.0"
description = "Sistema de gestión de productos y seguimiento de visitas"
authors = ["Tu Nombre <tu.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.2"
django-ninja = "1.3.0"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
celery = "^5.3.6"
python-jose = "^3.3.0"
sendgrid = "^6.11.0"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.0"
uvicorn = "^0.27.0"
django-redis = "^5.4.0"
pydantic = {extras = ["email"], version = "^2.6.0"}

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.0.0"
pytest = "^8.0.0"
pytest-django = "^4.8.0"
pytest-cov = "^4.1.0"
mypy = "^1.7.1"
pre-commit = "^3.5.0"
isort = "^5.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "product_watch.settings"
python_files = ["test_*.py"]
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
addopts = "--cov=src --cov-report=html --cov-fail-under=90"